requirements.txt
alembic==1.15.2
google_search_results==2.4.2
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.58
langchain-openai==0.3.16
langchain-text-splitters==0.3.8
langgraph==0.3.31
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.69
langsmith==0.3.42
openai==1.77.0
psycopg2-binary==2.9.10
pgvector==0.4.1
python-dotenv==1.1.0
SQLAlchemy==2.0.40
streamlit==1.45.0

